linuxkpi xa_*: Use uintptr_t instead of unsigned long.

+++ sys/compat/linuxkpi/common/include/linux/xarray.h
@@ -124,7 +124,7 @@ xa_init(struct xarray *xa)
 static inline void *
 xa_mk_value(unsigned long v)
 {
-	unsigned long r = (v << 1) | 1;
+	uintptr_t r = (v << 1) | 1;
 
 	return ((void *)r);
 }
@@ -132,7 +132,7 @@ xa_mk_value(unsigned long v)
 static inline bool
 xa_is_value(const void *e)
 {
-	unsigned long v = (unsigned long)e;
+	uintptr_t v = (uintptr_t)e;
 
 	return (v & 1);
 }
@@ -140,8 +140,17 @@ xa_is_value(const void *e)
 static inline unsigned long
 xa_to_value(const void *e)
 {
-	unsigned long v = (unsigned long)e;
+	uintptr_t v = (uintptr_t)e;
 
 	return (v >> 1);
 }
 #endif		/* _LINUXKPI_LINUX_XARRAY_H_ */
+// CHERI CHANGES START
+// {
+//   "updated": 20230424,
+//   "target_type": "kernel",
+//   "changes_purecap": [
+//     "pointer_as_integer"
+//   ]
+// }
+// CHERI CHANGES END
